Moved InventoryHandler
